refactor(rust): satisfy native release robustness gate
