+++ zeroshot-rust/src/cluster_ledger.rs
@@ -1,34 +1,29 @@
-use std::error::Error;
-use std::fmt;
 use std::sync::{Arc, Mutex};
 
-use serde::de::DeserializeOwned;
-use serde::Serialize;
-
-use crate::fault::{
-    EngineFault, EvidenceClass, FaultContext, FaultFactory, FaultModule, ModuleEvidence,
-};
+use crate::fault::{EvidenceClass, FaultContext, FaultFactory, FaultModule, ModuleEvidence};
 use crate::observability::{NoopObservationSink, ObservationSink};
 
 pub mod adapters;
+mod commit;
+mod error;
 pub mod mutations;
 pub mod record;
 pub mod replay;
 pub mod store;
 
 pub use mutations::{
-    AdmissionRequest, CommitResult, DispatchAllocation, SafeFaultConsequence, SafeFaultResult,
-    SettlementResult,
+    AdmissionRequest, CommitResult, DispatchAllocation, EffectReconciliation, SafeFaultConsequence,
+    SafeFaultRecord, SafeFaultResult, SettlementResult,
 };
+pub use error::{LedgerError, LedgerErrorKind};
 pub use record::{CanonicalDigest, EffectId, ExecutionId, GenerationId, NodeInstanceId, RunSequence};
 pub use replay::{replay, ReplayState};
 pub use store::{LedgerStore, OwnerId, ResourceId};
 
-use record::{RecordPayload, StoredRecord};
+pub(crate) use commit::{CommitRequest, MutationIdentity, ReceiptExpectation};
+use error::map_store_error;
 use replay::ReplayError;
-use store::{
-    AppendBatch, AppendGuard, AppendOutcome, Fence, IdempotencyId, MutationReceipt, StoreError,
-};
+use store::{Fence, IdempotencyId, MutationReceipt, StoreError};
 
 #[derive(Clone)]
 pub struct ClusterLedger {
@@ -38,32 +33,22 @@ pub struct ClusterLedger {
     observations: Arc<dyn ObservationSink>,
 }
 
-pub(crate) struct MutationIdentity {
-    key: IdempotencyId,
-    method: &'static str,
-    fingerprint: [u8; 32],
+#[derive(Clone)]
+pub struct LedgerAccess {
+    owner: OwnerId,
+    fence_ttl_ms: u64,
+    observations: Arc<dyn ObservationSink>,
 }
 
-impl MutationIdentity {
-    pub(crate) const fn new(
-        key: IdempotencyId,
-        method: &'static str,
-        fingerprint: [u8; 32],
-    ) -> Self {
+impl LedgerAccess {
+    #[must_use]
+    pub fn new(owner: OwnerId, fence_ttl_ms: u64, observations: Arc<dyn ObservationSink>) -> Self {
         Self {
-            key,
-            method,
-            fingerprint,
+            owner,
+            fence_ttl_ms,
+            observations,
         }
     }
-}
-
-struct PreparedCommit {
-    key: IdempotencyId,
-    method: &'static str,
-    fingerprint: [u8; 32],
-    receipt: MutationReceipt,
-    batch: AppendBatch,
 }
 
 impl ClusterLedger {
@@ -76,9 +61,7 @@ impl ClusterLedger {
         Self::create_with_observations(
             store,
             resource,
-            owner,
-            fence_ttl_ms,
-            Arc::new(NoopObservationSink),
+            LedgerAccess::new(owner, fence_ttl_ms, Arc::new(NoopObservationSink)),
         )
         .await
     }
@@ -86,10 +69,13 @@ impl ClusterLedger {
     pub async fn create_with_observations(
         store: Arc<dyn LedgerStore>,
         resource: ResourceId,
-        owner: OwnerId,
-        fence_ttl_ms: u64,
-        observations: Arc<dyn ObservationSink>,
+        access: LedgerAccess,
     ) -> Result<Self, LedgerError> {
+        let LedgerAccess {
+            owner,
+            fence_ttl_ms,
+            observations,
+        } = access;
         let (_, fence) = store
             .create_fenced(&resource, &owner, fence_ttl_ms)
             .await
@@ -113,9 +99,7 @@ impl ClusterLedger {
         Self::open_with_observations(
             store,
             resource,
-            owner,
-            fence_ttl_ms,
-            Arc::new(NoopObservationSink),
+            LedgerAccess::new(owner, fence_ttl_ms, Arc::new(NoopObservationSink)),
         )
         .await
     }
@@ -123,10 +107,13 @@ impl ClusterLedger {
     pub async fn open_with_observations(
         store: Arc<dyn LedgerStore>,
         resource: ResourceId,
-        owner: OwnerId,
-        fence_ttl_ms: u64,
-        observations: Arc<dyn ObservationSink>,
+        access: LedgerAccess,
     ) -> Result<Self, LedgerError> {
+        let LedgerAccess {
+            owner,
+            fence_ttl_ms,
+            observations,
+        } = access;
         store.open(&resource).await.map_err(|error| {
             map_store_error(observations.as_ref(), FaultContext::Recovery, error)
         })?;
@@ -220,223 +207,6 @@ impl ClusterLedger {
             .map_err(|error| self.store_error(FaultContext::Cleanup, error))
     }
 
-    pub(crate) async fn commit<T>(
-        &self,
-        context: FaultContext,
-        validated_state: &ReplayState,
-        mutation: MutationIdentity,
-        payloads: Vec<RecordPayload>,
-        response: &T,
-    ) -> Result<CommitResult<T>, LedgerError>
-    where
-        T: DeserializeOwned + Serialize,
-    {
-        self.commit_guarded(
-            context,
-            validated_state,
-            mutation,
-            payloads,
-            response,
-            AppendGuard::allow(),
-        )
-        .await
-    }
-
-    pub(crate) async fn commit_guarded<T>(
-        &self,
-        context: FaultContext,
-        validated_state: &ReplayState,
-        mutation: MutationIdentity,
-        payloads: Vec<RecordPayload>,
-        response: &T,
-        guard: AppendGuard,
-    ) -> Result<CommitResult<T>, LedgerError>
-    where
-        T: DeserializeOwned + Serialize,
-    {
-        let prepared =
-            self.prepare_commit(context, validated_state, mutation, payloads, response)?;
-        let outcome = self
-            .store
-            .compare_and_append_guarded(
-                &self.resource,
-                &self.fence(),
-                validated_state.position,
-                prepared.batch.clone(),
-                guard,
-            )
-            .await;
-        self.resolve_commit(context, prepared, outcome).await
-    }
-
-    fn prepare_commit<T>(
-        &self,
-        context: FaultContext,
-        validated_state: &ReplayState,
-        mutation: MutationIdentity,
-        payloads: Vec<RecordPayload>,
-        response: &T,
-    ) -> Result<PreparedCommit, LedgerError>
-    where
-        T: Serialize,
-    {
-        let MutationIdentity {
-            key,
-            method,
-            fingerprint,
-        } = mutation;
-        let (mut records, mut position, previous_hash) =
-            self.build_mutation_records(context, validated_state, payloads)?;
-        let encoded_response = serde_json::to_vec(response)
-            .map_err(|_| self.domain_error(context, LedgerErrorKind::Encoding))?;
-        position = position
-            .checked_add(1)
-            .map_err(|_| self.domain_error(context, LedgerErrorKind::BoundViolation))?;
-        let receipt = MutationReceipt {
-            idempotency_key: key.clone(),
-            method: method.to_owned(),
-            fingerprint,
-            response: encoded_response,
-            committed_position: position,
-        };
-        let receipt_record = StoredRecord::build(
-            self.resource.clone(),
-            position,
-            &RecordPayload::MutationReceipt {
-                receipt: receipt.clone(),
-            },
-            previous_hash,
-        )
-        .map_err(|_| self.domain_error(context, LedgerErrorKind::BoundViolation))?;
-        records.push(receipt_record);
-        let batch = AppendBatch::new(records, Some(receipt.clone()))
-            .map_err(|_| self.domain_error(context, LedgerErrorKind::BoundViolation))?;
-        Ok(PreparedCommit {
-            key,
-            method,
-            fingerprint,
-            receipt,
-            batch,
-        })
-    }
-
-    fn build_mutation_records(
-        &self,
-        context: FaultContext,
-        validated_state: &ReplayState,
-        payloads: Vec<RecordPayload>,
-    ) -> Result<(Vec<StoredRecord>, store::Position, [u8; 32]), LedgerError> {
-        let mut position = validated_state.position;
-        let mut previous_hash = validated_state.last_hash;
-        let mut records = Vec::with_capacity(payloads.len().saturating_add(1));
-        for payload in payloads {
-            position = position
-                .checked_add(1)
-                .map_err(|_| self.domain_error(context, LedgerErrorKind::BoundViolation))?;
-            let record =
-                StoredRecord::build(self.resource.clone(), position, &payload, previous_hash)
-                    .map_err(|_| self.domain_error(context, LedgerErrorKind::BoundViolation))?;
-            previous_hash = record.record_hash;
-            records.push(record);
-        }
-        Ok((records, position, previous_hash))
-    }
-
-    async fn resolve_commit<T>(
-        &self,
-        context: FaultContext,
-        prepared: PreparedCommit,
-        outcome: Result<AppendOutcome, StoreError>,
-    ) -> Result<CommitResult<T>, LedgerError>
-    where
-        T: DeserializeOwned,
-    {
-        let PreparedCommit {
-            key,
-            method,
-            fingerprint,
-            receipt,
-            ..
-        } = prepared;
-        match outcome {
-            Ok(AppendOutcome::Committed(committed)) if committed == receipt => {
-                self.decode_receipt(context, committed, method, fingerprint, false)
-            }
-            Ok(AppendOutcome::Committed(_)) => {
-                Err(self.domain_error(context, LedgerErrorKind::ReceiptCorrupt))
-            }
-            Ok(AppendOutcome::Replayed(committed)) => {
-                let recovered = self
-                    .validated_receipt(context, &key)
-                    .await?
-                    .ok_or_else(|| self.domain_error(context, LedgerErrorKind::ReceiptCorrupt))?;
-                if recovered != committed {
-                    return Err(self.domain_error(context, LedgerErrorKind::ReceiptCorrupt));
-                }
-                self.decode_receipt(context, recovered, method, fingerprint, true)
-            }
-            Ok(AppendOutcome::CommittedWithoutReceipt(_)) => {
-                Err(self.domain_error(context, LedgerErrorKind::ReceiptCorrupt))
-            }
-            Err(StoreError::FailureInjected(store::FailPoint::AfterCommitBeforeResponse)) => {
-                let recovered = self
-                    .validated_receipt(context, &key)
-                    .await?
-                    .ok_or_else(|| self.domain_error(context, LedgerErrorKind::ReceiptCorrupt))?;
-                self.decode_receipt(context, recovered, method, fingerprint, true)
-            }
-            Err(StoreError::IdempotencyConflict) => {
-                Err(self.domain_error(context, LedgerErrorKind::IdempotencyConflict))
-            }
-            Err(StoreError::PositionConflict { .. }) => {
-                Err(self.domain_error(context, LedgerErrorKind::InvalidLifecycle))
-            }
-            Err(error) => Err(self.store_error(context, error)),
-        }
-    }
-
-    fn decode_receipt<T>(
-        &self,
-        context: FaultContext,
-        receipt: MutationReceipt,
-        method: &str,
-        fingerprint: [u8; 32],
-        replayed: bool,
-    ) -> Result<CommitResult<T>, LedgerError>
-    where
-        T: DeserializeOwned,
-    {
-        if receipt.method != method || receipt.fingerprint != fingerprint {
-            return Err(self.domain_error(context, LedgerErrorKind::IdempotencyConflict));
-        }
-        let value = serde_json::from_slice(&receipt.response)
-            .map_err(|_| self.domain_error(context, LedgerErrorKind::ReceiptCorrupt))?;
-        Ok(CommitResult {
-            value,
-            position: receipt.committed_position,
-            replayed,
-        })
-    }
-
-    pub(crate) fn existing_receipt<T>(
-        &self,
-        context: FaultContext,
-        state: &ReplayState,
-        key: &IdempotencyId,
-        method: &str,
-        fingerprint: [u8; 32],
-    ) -> Result<Option<CommitResult<T>>, LedgerError>
-    where
-        T: DeserializeOwned,
-    {
-        state
-            .mutation_receipts
-            .get(key)
-            .cloned()
-            .map(|receipt| self.decode_receipt(context, receipt, method, fingerprint, true))
-            .transpose()
-    }
-
     pub(crate) async fn receipt(
         &self,
         context: FaultContext,
@@ -501,84 +271,3 @@ impl ClusterLedger {
         }
     }
 }
-
-fn map_store_error(
-    observations: &dyn ObservationSink,
-    context: FaultContext,
-    error: StoreError,
-) -> LedgerError {
-    let class = match error {
-        StoreError::BatchRecordBound
-        | StoreError::BatchByteBound
-        | StoreError::ReceiptTooLarge
-        | StoreError::InvalidLimit
-        | StoreError::PositionOverflow => EvidenceClass::ResourceExhausted,
-        StoreError::FenceHeld | StoreError::FenceExpired | StoreError::StaleFence => {
-            EvidenceClass::Unavailable
-        }
-        StoreError::ResourceNotFound | StoreError::ResourceExists => EvidenceClass::Unavailable,
-        StoreError::Storage | StoreError::FailureInjected(_) => EvidenceClass::Unavailable,
-        _ => EvidenceClass::IntegrityFailure,
-    };
-    let fault = FaultFactory::new(observations).create(ModuleEvidence::new(
-        FaultModule::Storage,
-        context,
-        class,
-    ));
-    LedgerError {
-        kind: LedgerErrorKind::Storage(error),
-        fault,
-    }
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub enum LedgerErrorKind {
-    Storage(StoreError),
-    Replay(ReplayError),
-    BoundViolation,
-    IdempotencyConflict,
-    InvalidLifecycle,
-    InvalidSettlement,
-    TerminalRequired,
-    ReceiptCorrupt,
-    Encoding,
-}
-
-pub struct LedgerError {
-    kind: LedgerErrorKind,
-    fault: EngineFault,
-}
-
-impl LedgerError {
-    #[must_use]
-    pub const fn kind(&self) -> &LedgerErrorKind {
-        &self.kind
-    }
-
-    #[must_use]
-    pub const fn fault(&self) -> &EngineFault {
-        &self.fault
-    }
-}
-
-impl fmt::Debug for LedgerError {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        formatter
-            .debug_struct("LedgerError")
-            .field("kind", &self.kind)
-            .field("fault_code", &self.fault.code())
-            .finish()
-    }
-}
-
-impl fmt::Display for LedgerError {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(
-            formatter,
-            "cluster ledger operation failed: {:?}",
-            self.kind
-        )
-    }
-}
-
-impl Error for LedgerError {}
