+++ zeroshot-rust/src/cluster_ledger/store/fake.rs
@@ -1,52 +1,23 @@
 use std::collections::BTreeMap;
-use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::{Arc, Mutex};
+
+mod append;
+mod clock;
 
 use async_trait::async_trait;
 use tokio::sync::Notify;
 
+use append::{apply_append, existing_outcome, take_failpoint, validate_append, validate_fence};
+pub use clock::ManualLedgerClock;
 use super::{
-    AppendBatch, AppendGuard, AppendOutcome, DiscoveryPage, FailPoint, Fence, IdempotencyId,
-    LedgerClock, LedgerStore, MutationReceipt, OwnerId, Position, PrefixSnapshot, ResourceId,
-    ResourceInfo, StoreError, MAX_DISCOVERY_PAGE,
+    completed_wait_position, fence_expiry, AppendOutcome, AppendRequest, DiscoveryPage, FailPoint,
+    Fence, IdempotencyId, LedgerClock, LedgerStore, MutationReceipt, OwnerId, Position,
+    PrefixSnapshot, ResourceId, ResourceInfo, StoreError, WaitProbe, MAX_DISCOVERY_PAGE,
 };
 use crate::cluster_ledger::record::{StoredRecord, MAX_RANGE_RECORDS};
 
-#[derive(Clone, Debug)]
-pub struct ManualLedgerClock {
-    now_ms: Arc<AtomicU64>,
-}
-
-impl ManualLedgerClock {
-    #[must_use]
-    pub fn new(now_ms: u64) -> Self {
-        Self {
-            now_ms: Arc::new(AtomicU64::new(now_ms)),
-        }
-    }
-
-    pub fn set(&self, now_ms: u64) {
-        self.now_ms.store(now_ms, Ordering::Release);
-    }
-
-    pub fn advance(&self, delta_ms: u64) -> Result<u64, StoreError> {
-        self.now_ms
-            .fetch_update(Ordering::AcqRel, Ordering::Acquire, |current| {
-                current.checked_add(delta_ms)
-            })
-            .map(|previous| previous + delta_ms)
-            .map_err(|_| StoreError::PositionOverflow)
-    }
-}
-
-impl LedgerClock for ManualLedgerClock {
-    fn now_ms(&self) -> u64 {
-        self.now_ms.load(Ordering::Acquire)
-    }
-}
-
 #[derive(Default)]
-struct ResourceData {
+pub(super) struct ResourceData {
     records: Vec<StoredRecord>,
     receipts: BTreeMap<IdempotencyId, MutationReceipt>,
     fence: Option<Fence>,
@@ -55,9 +26,16 @@ struct ResourceData {
 }
 
 #[derive(Default)]
-struct FakeState {
+pub(super) struct FakeState {
     resources: BTreeMap<ResourceId, ResourceData>,
-    next_failpoint: Option<FailPoint>,
+    pub(super) next_failpoint: Option<FailPoint>,
+}
+
+struct FenceRequest<'a> {
+    resource: &'a ResourceId,
+    owner: &'a OwnerId,
+    ttl_ms: u64,
+    epoch: u64,
 }
 
 #[derive(Clone)]
@@ -95,37 +73,13 @@ impl FakeLedgerStore {
             .next_failpoint = Some(point);
     }
 
-    fn take_failpoint(state: &mut FakeState, point: FailPoint) -> bool {
-        if state.next_failpoint == Some(point) {
-            state.next_failpoint = None;
-            true
-        } else {
-            false
-        }
-    }
-
-    fn validate_fence(
-        clock: &dyn LedgerClock,
-        data: &ResourceData,
-        fence: &Fence,
-    ) -> Result<(), StoreError> {
-        let current = data.fence.as_ref().ok_or(StoreError::StaleFence)?;
-        if current != fence {
-            return Err(StoreError::StaleFence);
-        }
-        if current.expires_at_ms <= clock.now_ms() {
-            return Err(StoreError::FenceExpired);
-        }
-        Ok(())
-    }
-
-    fn fence(
-        &self,
-        resource: &ResourceId,
-        owner: &OwnerId,
-        ttl_ms: u64,
-        epoch: u64,
-    ) -> Result<Fence, StoreError> {
+    fn fence(&self, request: FenceRequest<'_>) -> Result<Fence, StoreError> {
+        let FenceRequest {
+            resource,
+            owner,
+            ttl_ms,
+            epoch,
+        } = request;
         let expires_at_ms = self
             .clock
             .now_ms()
@@ -144,104 +98,48 @@ impl FakeLedgerStore {
         })
     }
 
-    fn append(
-        &self,
-        resource: &ResourceId,
-        fence: &Fence,
-        expected: Position,
-        batch: AppendBatch,
-        guard: &AppendGuard,
-    ) -> Result<AppendOutcome, StoreError> {
-        batch.validate()?;
+    fn append(&self, request: AppendRequest<'_>) -> Result<AppendOutcome, StoreError> {
+        request.batch.validate()?;
         let mut state = self
             .state
             .lock()
             .expect("fake ledger mutex must not be poisoned");
-        if Self::take_failpoint(&mut state, FailPoint::BeforeCommit) {
+        if take_failpoint(&mut state, FailPoint::BeforeCommit) {
             return Err(StoreError::FailureInjected(FailPoint::BeforeCommit));
         }
-        let data = state
-            .resources
-            .get_mut(resource)
-            .ok_or(StoreError::ResourceNotFound)?;
-        Self::validate_fence(self.clock.as_ref(), data, fence)?;
-        if let Some(outcome) = existing_outcome(data, &batch)? {
-            return Ok(outcome);
-        }
-        validate_append(data, resource, expected, &batch)?;
-        guard.check()?;
-        let outcome = apply_append(data, expected, batch)?;
-        if Self::take_failpoint(&mut state, FailPoint::AfterCommitBeforeResponse) {
+        let outcome = self.append_locked(&mut state, request)?;
+        if take_failpoint(&mut state, FailPoint::AfterCommitBeforeResponse) {
             return Err(StoreError::FailureInjected(
                 FailPoint::AfterCommitBeforeResponse,
             ));
         }
         Ok(outcome)
     }
-}
 
-fn existing_outcome(
-    data: &ResourceData,
-    batch: &AppendBatch,
-) -> Result<Option<AppendOutcome>, StoreError> {
-    let Some(receipt) = &batch.receipt else {
-        return Ok(None);
-    };
-    let Some(existing) = data.receipts.get(&receipt.idempotency_key) else {
-        return Ok(None);
-    };
-    if existing.method == receipt.method && existing.fingerprint == receipt.fingerprint {
-        Ok(Some(AppendOutcome::Replayed(existing.clone())))
-    } else {
-        Err(StoreError::IdempotencyConflict)
-    }
-}
-
-fn validate_append(
-    data: &ResourceData,
-    resource: &ResourceId,
-    expected: Position,
-    batch: &AppendBatch,
-) -> Result<(), StoreError> {
-    let actual = Position::new(
-        u64::try_from(data.records.len()).map_err(|_| StoreError::PositionOverflow)?,
-    )?;
-    if actual != expected {
-        return Err(StoreError::PositionConflict { expected, actual });
-    }
-    let committed_position = expected.checked_add(batch.records.len())?;
-    for (offset, record) in batch.records.iter().enumerate() {
-        let expected_sequence = expected.checked_add(offset + 1)?;
-        if record.resource != *resource || record.sequence != expected_sequence {
-            return Err(StoreError::Corrupt("append record identity"));
+    fn append_locked(
+        &self,
+        state: &mut FakeState,
+        request: AppendRequest<'_>,
+    ) -> Result<AppendOutcome, StoreError> {
+        let AppendRequest {
+            resource,
+            fence,
+            expected,
+            batch,
+            guard,
+        } = request;
+        let data = state
+            .resources
+            .get_mut(resource)
+            .ok_or(StoreError::ResourceNotFound)?;
+        validate_fence(self.clock.as_ref(), data, fence)?;
+        if let Some(outcome) = existing_outcome(data, &batch)? {
+            return Ok(outcome);
         }
+        validate_append(data, resource, expected, &batch)?;
+        guard.check()?;
+        apply_append(data, expected, batch)
     }
-    if batch
-        .receipt
-        .as_ref()
-        .is_some_and(|receipt| receipt.committed_position != committed_position)
-    {
-        return Err(StoreError::Corrupt("receipt position"));
-    }
-    Ok(())
-}
-
-fn apply_append(
-    data: &mut ResourceData,
-    expected: Position,
-    batch: AppendBatch,
-) -> Result<AppendOutcome, StoreError> {
-    let committed_position = expected.checked_add(batch.records.len())?;
-    data.records.extend(batch.records);
-    let outcome = if let Some(receipt) = batch.receipt {
-        data.receipts
-            .insert(receipt.idempotency_key.clone(), receipt.clone());
-        AppendOutcome::Committed(receipt)
-    } else {
-        AppendOutcome::CommittedWithoutReceipt(committed_position)
-    };
-    data.notify.notify_waiters();
-    Ok(outcome)
 }
 
 #[async_trait]
@@ -308,7 +206,12 @@ impl LedgerStore for FakeLedgerStore {
         owner: &OwnerId,
         ttl_ms: u64,
     ) -> Result<(ResourceInfo, Fence), StoreError> {
-        let fence = self.fence(resource, owner, ttl_ms, 1)?;
+        let fence = self.fence(FenceRequest {
+            resource,
+            owner,
+            ttl_ms,
+            epoch: 1,
+        })?;
         let mut state = self
             .state
             .lock()
@@ -394,16 +297,8 @@ impl LedgerStore for FakeLedgerStore {
     }
 
     async fn renew_fence(&self, fence: &Fence, ttl_ms: u64) -> Result<Fence, StoreError> {
-        if ttl_ms == 0 {
-            return Err(StoreError::FenceExpired);
-        }
         let now = self.clock.now_ms();
-        let expires_at_ms = now
-            .checked_add(ttl_ms)
-            .ok_or(StoreError::PositionOverflow)?;
-        if expires_at_ms > i64::MAX as u64 {
-            return Err(StoreError::PositionOverflow);
-        }
+        let expires_at_ms = fence_expiry(now, ttl_ms)?;
         let mut state = self
             .state
             .lock()
@@ -412,7 +307,7 @@ impl LedgerStore for FakeLedgerStore {
             .resources
             .get_mut(&fence.resource)
             .ok_or(StoreError::ResourceNotFound)?;
-        Self::validate_fence(self.clock.as_ref(), data, fence)?;
+        validate_fence(self.clock.as_ref(), data, fence)?;
         let renewed = Fence {
             expires_at_ms,
             ..fence.clone()
@@ -430,7 +325,7 @@ impl LedgerStore for FakeLedgerStore {
             .resources
             .get(&fence.resource)
             .ok_or(StoreError::ResourceNotFound)?;
-        Self::validate_fence(self.clock.as_ref(), data, fence)
+        validate_fence(self.clock.as_ref(), data, fence)
     }
 
     async fn read_prefix(
@@ -499,23 +394,9 @@ impl LedgerStore for FakeLedgerStore {
 
     async fn compare_and_append(
         &self,
-        resource: &ResourceId,
-        fence: &Fence,
-        expected: Position,
-        batch: AppendBatch,
+        request: AppendRequest<'_>,
     ) -> Result<AppendOutcome, StoreError> {
-        self.append(resource, fence, expected, batch, &AppendGuard::allow())
-    }
-
-    async fn compare_and_append_guarded(
-        &self,
-        resource: &ResourceId,
-        fence: &Fence,
-        expected: Position,
-        batch: AppendBatch,
-        guard: AppendGuard,
-    ) -> Result<AppendOutcome, StoreError> {
-        self.append(resource, fence, expected, batch, &guard)
+        self.append(request)
     }
 
     async fn lookup_receipt(
@@ -556,12 +437,15 @@ impl LedgerStore for FakeLedgerStore {
                 )?;
                 (position, notified)
             };
-            if position > after || self.clock.now_ms() >= deadline_ms {
-                return Ok(position);
-            }
-            let reread = self.open(resource).await?.position;
-            if reread > after {
-                return Ok(reread);
+            let probe = WaitProbe {
+                resource,
+                position,
+                after,
+                now_ms: self.clock.now_ms(),
+                deadline_ms,
+            };
+            if let Some(completed) = completed_wait_position(self, probe).await? {
+                return Ok(completed);
             }
             tokio::select! {
                 () = notified => {}
@@ -580,14 +464,14 @@ impl LedgerStore for FakeLedgerStore {
             .state
             .lock()
             .expect("fake ledger mutex must not be poisoned");
-        if Self::take_failpoint(&mut state, FailPoint::BeforeCommit) {
+        if take_failpoint(&mut state, FailPoint::BeforeCommit) {
             return Err(StoreError::FailureInjected(FailPoint::BeforeCommit));
         }
         let data = state
             .resources
             .get(resource)
             .ok_or(StoreError::ResourceNotFound)?;
-        Self::validate_fence(self.clock.as_ref(), data, fence)?;
+        validate_fence(self.clock.as_ref(), data, fence)?;
         let actual = Position::new(
             u64::try_from(data.records.len()).map_err(|_| StoreError::PositionOverflow)?,
         )?;
@@ -595,7 +479,7 @@ impl LedgerStore for FakeLedgerStore {
             return Err(StoreError::PositionConflict { expected, actual });
         }
         state.resources.remove(resource);
-        if Self::take_failpoint(&mut state, FailPoint::AfterCommitBeforeResponse) {
+        if take_failpoint(&mut state, FailPoint::AfterCommitBeforeResponse) {
             return Err(StoreError::FailureInjected(
                 FailPoint::AfterCommitBeforeResponse,
             ));
